Change the internal representation of definitions
Being able to access the entire definition Hash is something that is
not going to be possible anymore.

The library should be able to change the definitions without being
afraid of breaking external implementations. Developers should not
change the definitions directly, instead they should use #define to
append new definitions. That also means that the definition order
should not be relevant (GH-536).

This patch changes the internal representation of the definitions, and
hides them from the outside. It's not possible to access the definition
list only using #definitions and passing the type. Different types may
have different representations in the future.

The returned list is always a clone, so that the internal definition
list is not affected.