Add schema and metadata information to the definition file
Since the definition file is often used in third party applications, I
need a way to reference changes to the schema itself directly in the
file.

I adopted the approach where the will be always a _ element in the file
containing metadata, such as the schema version.

Consumers should always check if (and compare the version) before
processing the definition list.