Remove leading DOT from tlds
I know, this change may break some of the tools that are fetching this
file directly. I apologize in advance.

I need the format of this file to evolve, and I added schema
information in ef826e8895c9bc10103ba2e2f49ee843e413d456 to make sure in
the future consumers can check against it, and determine if any
breaking change occurred.

This is the first breaking change in several years.