Reduce string allocations and retentions (#580)
I'm trying to reduce these duplications as to reclaim a bit of memory. The whole thing rely on String#-@ which starting in MRI 2.5 will freeze and possibly intern the string.

On 2.4 and older (or some oldish alternative implementation) it will fallback to simply freezing the string, so no memory will be saved, but for the caller the string will still be frozen so that should prevent any kind of weirdness when upgrade the ruby version.